+++ woninet/core/subnet_enumerator.py
@@ -1,8 +1,8 @@
+import logging
 from typing import Dict
 from woninet.core.models import Device
-from woninet.utilities.logger import get_core_logger
 
-core_logger = get_core_logger()
+core_logger = logging.getLogger("core")
 
 
 class SubnetEnumerator: